Implement a nightly cleanup time
To improve performance, we'll offload the deleting of old cache files until the middle of the night instead of on every new request. Caches now have a 'cleanup_time' parameter which will be updated on a nightly basis when the current cleanup is finished.

This does mean that there may be more files than the  number, but after each cleanup, the  will be met again.